Python interface for multitarget bfs/dijkstra and extended SPSP
